requirements.txt
# Dependencias principales para monitoreo
requests>=2.31.0
ssl-checker>=1.0.0
python-whois>=0.9.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Para trabajar con fechas
python-dateutil>=2.8.2
pytz>=2023.3

# Para envío de emails
smtplib2>=0.2.1

# Para logging y debugging
loguru>=0.7.0

# Para trabajar con JSON y CSV
pandas>=2.0.0

# Para verificación de certificados SSL
cryptography>=41.0.0
pyOpenSSL>=23.2.0

# Para monitoreo web
urllib3>=2.0.0
beautifulsoup4>=4.12.0
